Fix numpy float64 crash in Gemini STST mean-reversion screener
sma_distance_pct was computed from pandas Series operations without
casting to native Python float. psycopg2 cannot serialize np.float64
and renders it literally as "np.float64(-10.0)" in the SQL, which
PostgreSQL interprets as schema.function — causing:

  ProgrammingError: schema "np" does not exist

Fix: wrap the pandas arithmetic in float() before round(), matching
the pattern used by all other fields in the signal dict.

+++ Gemini STST/app/mean_reversion.py
@@ -264,7 +264,7 @@ def run_reversion_screener(screen_date: date | None = None) -> dict:
 
         # SMA distance: how far below the 20-day SMA (rubber-band stretch)
         sma_20 = df["close"].rolling(20).mean().iloc[-1]
-        sma_distance_pct = round(((latest["close"] / sma_20) - 1.0) * 100, 1) if not pd.isna(sma_20) else 0.0
+        sma_distance_pct = round(float((latest["close"] / sma_20) - 1.0) * 100, 1) if not pd.isna(sma_20) else 0.0
 
         # ATR% for vol-scaled sizing
         atr_pct_val = round(float(latest["atr_pct"]), 1) if not pd.isna(latest["atr_pct"]) else 10.0
